+++ src/Geo/Traits/Addressable.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Enraiged\Geo\Traits;
+
+use Enraiged\Geo\Models\Address;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Database\Eloquent\Relations\MorphOne;
+
+trait Addressable
+{
+    /**
+     *  @return \Illuminate\Database\Eloquent\Relations\MorphOne
+     */
+    public function address(): MorphOne
+    {
+        return $this->morphOne(Address::class, 'addressable')
+            ->where('is_default', true);
+    }
+
+    /**
+     *  @return \Illuminate\Database\Eloquent\Relations\MorphMany
+     */
+    public function addresses(): MorphMany
+    {
+        return $this->morphMany(Address::class, 'addressable');
+    }
+}