+++ src/Users/Forms/Validation/Rules.php
@@ -2,6 +2,7 @@
 
 namespace Enraiged\Users\Forms\Validation;
 
+use Enraiged\Geo\Models\Country;
 use Enraiged\Passwords\Forms\Validation\PasswordRules;
 use Enraiged\Users\Enums\Roles;
 use Illuminate\Support\Str;
@@ -11,18 +12,45 @@ trait Rules
 {
     /** @var  array  the validation rules that apply to the request. */
     protected $rules = [
+        'birthday' => 'nullable|date',
+        'country_id' => 'nullable|int|exists:countries,id',
+        'is_active' => 'boolean',
         'name' => 'required|string|max:255',
+        'phone' => 'nullable|string|max:12',
         'salut' => 'nullable|string',
         'theme' => 'array',
-        'timezone' => 'nullable|string|max:100',
+        'timezone' => 'nullable|timezone',
     ];
+
+    /**
+     *  Assemble and return the email validation rule for the request.
+     *
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @param  \Enraiged\Users\Enums\Roles  $roles
+     *  @return array
+     */
+    protected function validateEmailRule($user, $roles)
+    {
+        $table = $user->getTable();
+
+        return [
+            'required',
+            'string',
+            'lowercase',
+            'email',
+            'max:255',
+            Rule::unique($table, 'email')->ignore($user->id),
+        ];
+    }
 
     /**
      *  Assemble and return the locale validation rule for the request.
      *
+     *  @param  \Enraiged\Users\Models\User  $user
+     *  @param  \Enraiged\Users\Enums\Roles  $roles
      *  @return array
      */
-    protected function validateLocaleRule()
+    protected function validateLocaleRule($user, $roles)
     {
         $locales = collect(config('enraiged.locales'))
             ->keys()
@@ -39,32 +67,35 @@ trait Rules
     }
 
     /**
-     *  Assemble and return the email validation rule for the request.
+     *  Assemble and return the role_id validation rule for the request.
      *
      *  @param  \Enraiged\Users\Models\User  $user
-     *  @return array
+     *  @param  \Enraiged\Users\Enums\Roles  $roles
+     *  @return array|null
      */
-    protected function validateEmailRule($user)
+    protected function validateRoleIdRule($user, $roles)
     {
-        $table = $user->getTable();
+        $admin = config('auth.providers.roles.admin', 'Administrator');
 
-        return [
-            'required',
-            'string',
-            'lowercase',
-            'email',
-            'max:255',
-            Rule::unique($table, 'email')->ignore($user->id),
-        ];
+        if ($this->user()->role->is($roles::find($admin))) {
+            return [
+                'required',
+                'int',
+                'in:'.collect($roles::ids())->join(','),
+            ];
+        }
+
+        return null;
     }
 
     /**
      *  Assemble and return the username validation rule for the request.
      *
      *  @param  \Enraiged\Users\Models\User  $user
+     *  @param  \Enraiged\Users\Enums\Roles  $roles
      *  @return array
      */
-    protected function validateUsernameRule($user)
+    protected function validateUsernameRule($user, $roles)
     {
         $rules = ['nullable', 'string'];
 
@@ -102,10 +133,20 @@ trait Rules
             $user = $this->user
                 ? $model::findOrFail($this->user)
                 : new $model;
+        }
 
+        $rules = [
+            'password' => [
+                $user->exists ? 'sometimes' : 'required',
+                'confirmed',
+                new PasswordRules],
+        ];
+
+        if ($this->user()->role->is($roles::find('Administrator'))) {
             $rules = [
-                'password' => ['required', 'sometimes', 'confirmed', new PasswordRules],
-                'role_id' => ['required', 'sometimes', 'in:'.collect($roles::ids())->join(',')],
+                ...$rules,
+                'is_hidden' => 'boolean',
+                'is_protected' => 'boolean',
             ];
         }
 
@@ -116,10 +157,14 @@ trait Rules
             } else {
                 $method = 'validate'.ucwords(Str::camel($index)).'Rule';
 
-                if (method_exists($this, $method)) {
-                    $rules[$index] = $this->{$method}($user);
+                if (method_exists($this, $method) && $rule = $this->{$method}($user, $roles)) {
+                    $rules[$index] = $rule;
                 }
             }
+        }
+
+        if ($this->filled('country_id') && $country = Country::find($this->get('country_id'))) {
+            $rules['timezone'] = "{$rules['timezone']}:per_country,{$country->code}";
         }
 
         return $this->attribute